Removed some compiler warnings.

+++ examples/cpo_hdc.f90
@@ -11,7 +11,7 @@ program cpo_hdc
 contains
 
     subroutine test_cpos(equilibriumin, distsourceout)
-        integer :: i
+        integer(kind=8) :: i
         !UAL ! Always describe cpo as array 
         !UAL ! In case of time slice, the size of the input cpo is 1
         !UAL type (type_equilibrium),pointer :: equilibriumin(:)
